# Conflicts:
#	erp-server/erp-server-wms/src/main/java/com/erp/server/wms/schedule/DataRecoveryJob.java

+++ erp-model/erp-model-wms/src/main/java/com/erp/model/wms/dto/QcProductDTO.java
@@ -117,7 +117,7 @@ public class QcProductDTO {
         /**
          * 整箱数量
          */
-        private Integer boxQty;
+        private BigDecimal boxQty;
 
         /**
          * 产品净重
